pyproject.toml
[project]
name = "translator-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "audioop-lts>=0.2.2",
    "openai>=2.7.2",
    "py-cord>=2.6.1",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
]
